user stats for forum posts per course
--HG--
branch : Blog

+++ modules/forum/viewtopic.php
@@ -58,7 +58,7 @@ if (isset($_GET['topic'])) {
 }
 if (isset($_GET['post_id'])) {//needed to find post page for anchors
     $post_id = intval($_GET['post_id']);
-    $myrow = Database::get()->querySingle("SELECT f.id, f.name, p.post_time FROM forum f, forum_topic t, forum_post p
+    $myrow = Database::get()->querySingle("SELECT f.id, f.name, p.post_time, p.poster_id FROM forum f, forum_topic t, forum_post p
             WHERE f.id = ?d
             AND t.id = ?d
             AND p.id = ?d
@@ -83,18 +83,15 @@ $forum_name = $myrow->name;
 $forum = $myrow->id;
 $total = get_total_posts($topic);
 
-if (isset($_GET['delete']) && $is_editor) {
+if (isset($_GET['delete']) && isset($post_id) && $is_editor) {
     $idx = new Indexer();
     $ftdx = new ForumTopicIndexer($idx);
     $fpdx = new ForumPostIndexer($idx);
 
-    $post_id = intval($_GET['post_id']);
     $last_post_in_thread = get_last_post($topic);
 
-    $myrow = Database::get()->querySingle("SELECT post_time FROM forum_post
-                            WHERE id = ?d", $post_id);
-
     $this_post_time = $myrow->post_time;
+    $this_post_author = $myrow->poster_id;
 
     Database::get()->query("DELETE FROM forum_post WHERE id = ?d", $post_id);
     $fpdx->remove($post_id);
@@ -102,8 +99,17 @@ if (isset($_GET['delete']) && $is_editor) {
     //orphan replies get -1 to parent_post_id
     Database::get()->query("UPDATE forum_post SET parent_post_id = -1 WHERE parent_post_id = ?d", $post_id);
 
+    $forum_user_stats = Database::get()->querySingle("SELECT COUNT(*) as c FROM forum_post
+                        INNER JOIN forum_topic ON forum_post.topic_id = forum_topic.id
+                        INNER JOIN forum ON forum.id = forum_topic.forum_id
+                        WHERE forum_post.poster_id = ?d AND forum.course_id = ?d", $this_post_author, $course_id);
+    Database::get()->query("DELETE FROM forum_user_stats WHERE user_id = ?d AND course_id = ?d", $this_post_author, $course_id);
+    if ($forum_user_stats->c != 0) {
+        Database::get()->query("INSERT INTO forum_user_stats (user_id, num_posts, course_id) VALUES (?d,?d,?d)", $this_post_author, $forum_user_stats->c, $course_id);
+    }
+    
     if ($total == 1) { // if exists one post in topic
-        Database::get()->query("DELETE FROM forum_topic WHERE id = ?d AND forum_id = ", $topic, $forum);
+        Database::get()->query("DELETE FROM forum_topic WHERE id = ?d AND forum_id = ?d", $topic, $forum);
         $ftdx->remove($topic);
         Database::get()->query("UPDATE forum SET num_topics = 0,
                             num_posts = 0
@@ -254,13 +260,20 @@ if (isset($_GET['all'])) {
 }
 
 $count = 0;
+$user_stats = array();
 foreach ($result as $myrow) {
     if ($count % 2 == 1) {
         $tool_content .= "<tr class='odd'>";
     } else {
         $tool_content .= "<tr class='even'>";
     }
-    $tool_content .= "<td valign='top'>" . display_user($myrow->poster_id) . "</td>";
+    
+    if (!isset($user_stats[$myrow->poster_id])) {
+        $user_num_posts = Database::get()->querySingle("SELECT num_posts FROM forum_user_stats WHERE user_id = ?d AND course_id = ?d", $myrow->poster_id, $course_id);
+        $user_stats[$myrow->poster_id] = $user_num_posts->num_posts;
+    }
+    
+    $tool_content .= "<td valign='top'>" . display_user($myrow->poster_id) . "<br/>".$user_stats[$myrow->poster_id]." $langMessages</td>";
     $message = $myrow->post_text;
     // support for math symbols
     $message = mathfilter($message, 12, "../../courses/mathimg/");
